requirements.txt
aiogram==3.1.1
Pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
replicate==0.18.1
pip==23.3.1
